+++ CHANGELOG.md
@@ -9,6 +9,8 @@ IMPROVEMENTS:
 - Route entry supports Nat Gateway [GH-713]
 - Modify db account returning update to read after creating [GH-711]
 - Improve cdn testcase [GH-708]
+- Apply wraperror to security_group, security_group_rule, vswitch, disk [GH-707]
+- Improve cdn testcase [GH-705]
 - Add notes for datahub and improve its testcase [GH-704]
 - Improve security_group_rule resource and data source testcases [GH-703]
 - Improve kvstore backup policy [GH-701]